add if codition in the controller that record is in database or not

+++ app/Http/Controllers/Api/countryController.php
@@ -3,40 +3,63 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
 use App\Model\Api\country;
-
+use Illuminate\Http\Request;
 
 class countryController extends Controller
 {
     public function getCounty()
     {
 
-        return response()->json(country::get(),200);
+        return response()->json(country::get(), 200);
     }
     public function getCountryById($id)
     {
-        return response()->json(country::find($id),200);
+
+        $country = country::find($id);
+        if (is_null($country)) {
+            return response()->json(['message', ' This id data record not found'], 400);
+        } else {
+            return response()->json($country, 200);
+        }
+
     }
     public function saveCountydata(Request $request)
     {
 
-        $country=country::create($request->all());
-        return response()->json($country,201);
-    }
-    public function editCountrydata(Request $request,country $country)
-
-    {
-        $country->update($request->all());
-        return response()->json($country,200);
-
-
+        $country = country::create($request->all());
+        return response()->json($country, 201);
     }
 
-    public function deleteCountrydata(Request $request, country $country)
+    public function editCountrydata(Request $request, $id)
     {
+        $country = country::find($id);
+
+        if (is_null($country)) {
+            return response()->json(['message', ' This id data record not found'], 400);
+        } else {
+            $country->update($request->all());
+
+            return response()->json($country, 200);
+
+        }
+    }
+
+    public function deleteCountrydata(Request $request, $id)
+    {
+
+        $country=country::find($id);
+        if(is_null($country))
+        {
+
+            return response()->json(['message','this id data record is not found'],400);
+        }
+        else 
+        {
 
         $country->delete();
-        return response()->json(null,204);
+
+        return response()->json(null, 204);
+            }
     }
 }
